Refactored tests and added time mocking
