+++ pyproject.toml
@@ -46,6 +46,7 @@ test = [
     "pytest ==5.4.3",
     "pytest-cov ==2.10.0",
     "pytest-asyncio >=0.14.0,<0.15.0",
+    "pytest-mock ==3.6",
     "mypy ==0.812",
     "flake8 >=3.8.3,<4.0.0",
     "black ==20.8b1",
@@ -57,6 +58,7 @@ test = [
     "peewee >=3.13.3,<4.0.0",
     "databases[sqlite] >=0.3.2,<0.4.0",
     "orjson >=3.2.1,<4.0.0",
+    "pyjwt >=2.0.0,<3.0.0",
     "async_exit_stack >=1.0.1,<2.0.0",
     "async_generator >=1.10,<2.0.0",
     "python-multipart >=0.0.5,<0.0.6",
@@ -89,6 +91,7 @@ all = [
     "graphene >=2.1.8,<3.0.0",
     "ujson >=3.0.0,<4.0.0",
     "orjson >=3.2.1,<4.0.0",
+    "pyjwt >=2.0.0,<3.0.0",
     "email_validator >=1.1.1,<2.0.0",
     "uvicorn[standard] >=0.12.0,<0.14.0",
     "async_exit_stack >=1.0.1,<2.0.0",
