+++ tests/test_security_jwt_general.py
@@ -1,15 +1,14 @@
-from typing import Optional
+import datetime
 
 from fastapi import FastAPI, Security
 from fastapi.security.jwt import JwtAuth, JwtAuthCredentials, JwtAuthRefresh
 from fastapi.testclient import TestClient
+from pytest_mock import MockerFixture
 
 app = FastAPI()
 
-access_security = JwtAuth(secret_key="secret_key", places={"header"}, auto_error=False)
-refresh_security = JwtAuthRefresh(
-    secret_key="secret_key", places={"header"}, auto_error=False
-)
+access_security = JwtAuth(secret_key="secret_key")
+refresh_security = JwtAuthRefresh(secret_key="secret_key")
 
 
 @app.post("/auth")
@@ -23,10 +22,7 @@ def auth():
 
 
 @app.post("/refresh")
-def refresh(credentials: Optional[JwtAuthCredentials] = Security(refresh_security)):
-    if credentials is None:
-        return {"msg": "Create an account first"}
-
+def refresh(credentials: JwtAuthCredentials = Security(refresh_security)):
     access_token = refresh_security.create_access_token(subject=credentials.subject)
     refresh_token = refresh_security.create_refresh_token(subject=credentials.subject)
 
@@ -34,12 +30,18 @@ def refresh(credentials: Optional[JwtAuthCredentials] = Security(refresh_securit
 
 
 @app.get("/users/me")
-def read_current_user(
-    credentials: Optional[JwtAuthCredentials] = Security(access_security),
-):
-    if credentials is None:
-        return {"msg": "Create an account first"}
+def read_current_user(credentials: JwtAuthCredentials = Security(access_security)):
     return {"username": credentials["username"], "role": credentials["role"]}
+
+
+class _FakeDateTime(datetime.datetime):  # pragma: no cover
+    @staticmethod
+    def now(**kwargs):
+        return datetime.datetime.now() + datetime.timedelta(days=42)
+
+    @staticmethod
+    def utcnow(**kwargs):
+        return datetime.datetime.utcnow() + datetime.timedelta(days=42)
 
 
 client = TestClient(app)
@@ -112,12 +114,7 @@ def test_openapi_schema():
     assert response.json() == openapi_schema
 
 
-def test_security_jwt_auth():
-    response = client.post("/auth")
-    assert response.status_code == 200, response.text
-
-
-def test_security_jwt_access_bearer():
+def test_security_jwt_access_token():
     access_token = client.post("/auth").json()["access_token"]
 
     response = client.get(
@@ -127,27 +124,47 @@ def test_security_jwt_access_bearer():
     assert response.json() == {"username": "username", "role": "user"}
 
 
-def test_security_jwt_access_bearer_wrong():
+def test_security_jwt_access_token_wrong():
     response = client.get(
         "/users/me", headers={"Authorization": "Bearer wrong_access_token"}
     )
-    assert response.status_code == 200, response.text
-    assert response.json() == {"msg": "Create an account first"}
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith("Wrong token:")
+
+    response = client.get(
+        "/users/me", headers={"Authorization": "Bearer wrong.access.token"}
+    )
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith("Wrong token:")
 
 
-def test_security_jwt_access_bearer_no_credentials():
-    response = client.get("/users/me")
-    assert response.status_code == 200, response.text
-    assert response.json() == {"msg": "Create an account first"}
+def test_security_jwt_access_token_changed():
+    access_token = client.post("/auth").json()["access_token"]
+
+    access_token = access_token.split(".")[0] + ".wrong." + access_token.split(".")[-1]
+
+    response = client.get(
+        "/users/me", headers={"Authorization": f"Bearer {access_token}"}
+    )
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith("Wrong token:")
 
 
-def test_security_jwt_access_bearer_incorrect_scheme_credentials():
-    response = client.get("/users/me", headers={"Authorization": "Basic notreally"})
-    assert response.status_code == 200, response.text
-    assert response.json() == {"msg": "Create an account first"}
+def test_security_jwt_access_token_timeout(mocker: MockerFixture):
+    access_token = client.post("/auth").json()["access_token"]
+
+    mocker.patch("jwt.api_jwt.datetime", _FakeDateTime)  # 42 days left
+
+    response = client.get(
+        "/users/me", headers={"Authorization": f"Bearer {access_token}"}
+    )
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith(
+        "Token time expired: Signature has expired"
+    )
 
 
-def test_security_jwt_refresh_bearer():
+def test_security_jwt_refresh_token():
     refresh_token = client.post("/auth").json()["refresh_token"]
 
     response = client.post(
@@ -156,21 +173,55 @@ def test_security_jwt_refresh_bearer():
     assert response.status_code == 200, response.text
 
 
-def test_security_jwt_refresh_bearer_wrong():
+def test_security_jwt_refresh_token_wrong():
     response = client.post(
         "/refresh", headers={"Authorization": "Bearer wrong_refresh_token"}
     )
-    assert response.status_code == 200, response.text
-    assert response.json() == {"msg": "Create an account first"}
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith("Wrong token:")
+
+    response = client.post(
+        "/refresh", headers={"Authorization": "Bearer wrong.refresh.token"}
+    )
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith("Wrong token:")
 
 
-def test_security_jwt_refresh_bearer_no_credentials():
-    response = client.post("/refresh")
-    assert response.status_code == 200, response.text
-    assert response.json() == {"msg": "Create an account first"}
+def test_security_jwt_refresh_token_using_access_token():
+    tokens = client.post("/auth").json()
+    access_token, refresh_token = tokens["access_token"], tokens["refresh_token"]
+    assert access_token != refresh_token
+
+    response = client.post(
+        "/refresh", headers={"Authorization": f"Bearer {access_token}"}
+    )
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith("Wrong token: 'type' is not 'refresh'")
 
 
-def test_security_jwt_refresh_bearer_incorrect_scheme_credentials():
-    response = client.post("/refresh", headers={"Authorization": "Basic notreally"})
-    assert response.status_code == 200, response.text
-    assert response.json() == {"msg": "Create an account first"}
+def test_security_jwt_refresh_token_changed():
+    refresh_token = client.post("/auth").json()["refresh_token"]
+
+    refresh_token = (
+        refresh_token.split(".")[0] + ".wrong." + refresh_token.split(".")[-1]
+    )
+
+    response = client.post(
+        "/refresh", headers={"Authorization": f"Bearer {refresh_token}"}
+    )
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith("Wrong token:")
+
+
+def test_security_jwt_refresh_token_timeout(mocker: MockerFixture):
+    refresh_token = client.post("/auth").json()["refresh_token"]
+
+    mocker.patch("jwt.api_jwt.datetime", _FakeDateTime)  # 42 days left
+
+    response = client.post(
+        "/refresh", headers={"Authorization": f"Bearer {refresh_token}"}
+    )
+    assert response.status_code == 401, response.text
+    assert response.json()["detail"].startswith(
+        "Token time expired: Signature has expired"
+    )
